gcp/requirements.txt

google-auth
google-auth-oauthlib
requests
